+++ tests/data/stubs/stub-doas.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+
+import json
+import os
+import sys
+
+os.environ['ORIGINAL_ARGV'] = json.dumps(sys.argv)
+os.environ['THIS_IS_STUB_DOAS'] = '1'
+os.execv(sys.executable, sys.argv[sys.argv.index('--') + 1:])